# api/docs/swagger.yaml
basePath: /
definitions:
...
  health.AddMedicalReportReq:
    properties:
      attachments:
        items:
          $ref: '#/definitions/health.Attachment'
        type: array
      description:
        type: string
      doctor_id:
        type: string
      record_date:
        type: string
      record_type:
        type: string
      user_id:
        type: string
    type: object
  health.AddWearableDataReq:
    properties:
      data_type:
        type: string
      data_value:
        type: string
      device_type:
        type: string
      user_id:
        type: string
    type: object
  health.Attachment:
    properties:
      file_url:
        type: string
    type: object
...
  health.GetMedicalReportByIdRes:
    properties:
      attachments:
        items:
          $ref: '#/definitions/health.Attachment'
        type: array
      description:
        type: string
      doctor_id:
        type: string
      doctor_name:
        type: string
      email:
        type: string
      id:
        type: string
      record_date:
        type: string
      record_type:
        type: string
      user_id:
        type: string
    type: object
  health.GetWearableDataByIdRes:
    properties:
      get_warable_id:
        items:
          $ref: '#/definitions/health.WearableById'
        type: array
    type: object
...
  health.UpdateMedicalReportReq:
    properties:
      attachments:
        items:
          $ref: '#/definitions/health.Attachment'
        type: array
      description:
        type: string
      doctor_id:
        type: string
      id:
        type: string
      record_type:
        type: string
    type: object
  health.UpdateWearableDataReq:
    properties:
      data_type:
        type: string
      data_value:
        type: string
      device_type:
        type: string
      id:
        type: string
    type: object
  health.WearableById:
    properties:
      data_type:
        type: string
      data_value:
        type: string
      device_type:
        type: string
      id:
        type: string
      recorded_timestamp:
        type: string
      user_id:
        type: string
    type: object
  models.Error:
    properties:
      message:
        type: string
    type: object
...
  models.MedicalReport:
    properties:
      attachments:
        items:
          type: string
        type: array
      description:
        type: string
      doctor_name:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      record_type:
        type: string
      recorded_date:
        type: string
    type: object
  models.Success:
    properties:
      message:
        type: string
    type: object
...
  models.Warable:
    properties:
      data_type:
        type: string
      data_value:
        type: string
      device_type:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      recorded_timestamp:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: api gateway service
  title: Api Getaway
  version: "1.0"
paths:
...
  /api/medicalReport/add:
    post:
      description: Adds a medical report for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body for adding a medical report
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/health.AddMedicalReportReq'
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Success'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Add medical report
      tags:
      - MedicalReport
  /api/medicalReport/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a specific medical report by its ID
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete medical report
      tags:
      - MedicalReport
  /api/medicalReport/get/{user_id}:
    get:
      description: Retrieves all medical reports for a user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.MedicalReport'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get medical reports
      tags:
      - MedicalReport
  /api/medicalReport/getById/{id}:
    get:
      description: Retrieves a specific medical report by its ID
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/health.GetMedicalReportByIdRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get medical report by ID
      tags:
      - MedicalReport
  /api/medicalReport/update:
    put:
      consumes:
      - application/json
      description: Updates a specific medical report
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body for updating a medical report
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/health.UpdateMedicalReportReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Success'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Update medical report
      tags:
      - MedicalReport
...
  /api/wearable/add:
    post:
      consumes:
      - application/json
      description: Adds wearable data for a user
      parameters:
      - description: Request body for adding wearable data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/health.AddWearableDataReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Success'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Add wearable data
      tags:
      - WearableData
  /api/wearable/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a specific wearable data entry by its ID
      parameters:
      - description: Wearable Data ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete wearable data
      tags:
      - WearableData
  /api/wearable/get/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieves all wearable data for a user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Warable'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get wearable data
      tags:
      - WearableData
  /api/wearable/getById/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a specific wearable data entry by its ID
      parameters:
      - description: Wearable Data ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/health.GetWearableDataByIdRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get wearable data by ID
      tags:
      - WearableData
  /api/wearable/update:
    put:
      consumes:
      - application/json
      description: Updates a specific wearable data entry
      parameters:
      - description: Wearable Data ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body for updating wearable data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/health.UpdateWearableDataReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Success'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Update wearable data
      tags:
      - WearableData
schemes:
- http
swagger: "2.0"